移動のテスト修正

+++ haruhi/app/Http/Controllers/MoveController.php
@@ -5,25 +5,44 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Service\MoveService;
 use App\Exceptions\InvalidParameterException;
+use App\Exceptions\NotFoundException;
 use App\Util\DateUtil;
 
 class MoveController extends Controller
 {
+    private const MOVE_ID = 1;
+
     // @todo 入力値をそのまま入れているのでバリデーションを入れる
     public function index(string $attributeName)
     {
+        // 要素名が想定内かどうか
+        if ($attributeName !== 'purposes' && $attributeName !== 'places') {
+            throw new InvalidParameterException('Attribute name is wrong.');
+        }
+
         $moveService = new MoveService($attributeName);
         return $moveService->index();
     }
 
     public function show(string $attributeName, int $id)
     {
+        // 要素名が想定内かどうか
+        if ($attributeName !== 'purposes' && $attributeName !== 'places') {
+            throw new InvalidParameterException('Attribute name is wrong.');
+        }
+
         if (is_null($id) || !is_numeric($id)) {
             throw new InvalidParameterException('Move id is null.');
         }
 
         $moveService = new MoveService($attributeName);
-        return $moveService->show($id);
+        $moves = $moveService->show($id);
+
+        if (count($moves) === 0) {
+            throw new NotFoundException();
+        }
+
+        return $moves[0];
     }
 
     public function store(Request $request, string $attributeName)
@@ -37,11 +56,17 @@ class MoveController extends Controller
 
         self::validation($move);
         $moveService = new MoveService($attributeName);
-        $moveService->store($move);
+        $id = $moveService->store($move);
+        return ['id' => $id];
     }
 
     public function update(Request $request, string $attributeName, int $id)
     {
+        // 要素名が想定内かどうか
+        if ($attributeName !== 'purposes' && $attributeName !== 'places') {
+            throw new InvalidParameterException('Attribute name is wrong.');
+        }
+
         $move = [];
         $move['item'] = (string)$request->input('item');
         $move['amount'] = (int)$request->input('amount');
@@ -55,12 +80,28 @@ class MoveController extends Controller
 
         self::validation($move);
         $moveService = new MoveService($attributeName);
+
+        $moves = $moveService->show($id);
+        if (count($moves) === 0) {
+            throw new NotFoundException();
+        }
+
         $moveService->update($move, $id);
     }
 
     public function destroy(string $attributeName, int $id)
     {
+        // 要素名が想定内かどうか
+        if ($attributeName !== 'purposes' && $attributeName !== 'places') {
+            throw new InvalidParameterException('Attribute name is wrong.');
+        }
+
         $moveService = new MoveService($attributeName);
+
+        $moves = $moveService->show($id);
+        if (count($moves) === 0) {
+            throw new NotFoundException();
+        }
         $moveService->destroy($id);
     }
 
@@ -74,6 +115,17 @@ class MoveController extends Controller
         // 項目は空文字ではないとダメ
         if ($move['item'] === '') {
             throw new InvalidParameterException('Item is empty.');
+        }
+
+        // 入力されていないとダメ
+        if ($move['beforeId'] === 0 || $move['afterId'] === 0) {
+            throw new InvalidParameterException('Before id and after id is null.');
+        }
+
+        // 移動前後で移動IDを使っちゃダメ
+        // @todo purpose, place のモデルの定数に変える
+        if ($move['beforeId'] === self::MOVE_ID || $move['afterId'] === self::MOVE_ID) {
+            throw new InvalidParameterException('Before id or after id is move id.');
         }
 
         // 移動していないとダメ
